requirements.txt
django~=1.10
django-braces
